switch to better-sqlite3

+++ runtests.ps1
@@ -1,3 +1,6 @@
-Remove-Item .\errorlog.txt
+if (Test-Path -Path .\errorlog.txt) {
+    Remove-Item .\errorlog.txt
+}
+
 & sqlite3 -init "./db/initialize.sql" "./db/test.db" .quit
 node .\chatbottest.js
